+++ tests/lightning_tensor/test_var.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 """
-Tests for process and execute (variance calculation).
+Tests for the variance calculation on the LightningTensor device.
 """
 import numpy as np
 import pennylane as qml